rename according to the file's name convention
